+++ geofileops/util/_sqlite_util.py
@@ -87,6 +87,53 @@ def create_new_spatialdb(path: Path, crs_epsg: Optional[int] = None):
         conn.close()
 
 
+def get_columns(
+    sql_stmt: str,
+    input1_path: Path,
+    input2_path: Optional[Path] = None,
+    use_spatialite: bool = True,
+) -> List[str]:
+    input1_databasename = "main"
+    conn = sqlite3.connect(str(input1_path), uri=True)
+    sql = None
+    try:
+        if use_spatialite is True:
+            load_spatialite(conn)
+            if input1_path.suffix.lower() == ".gpkg":
+                sql = "SELECT EnableGpkgMode();"
+                conn.execute(sql)
+
+        # If input2 isn't the same database input1, attach to it
+        input2_databasename = None
+        if input2_path is not None:
+            if input2_path == input1_path:
+                input2_databasename = input1_databasename
+            else:
+                input2_databasename = "input2"
+                sql = f"ATTACH DATABASE ? AS {input2_databasename}"
+                # dbSpec = (f"file:{input2_path.resolve().as_posix()}?mode=ro",)
+                dbSpec = (str(input2_path),)
+                conn.execute(sql, dbSpec)
+        # Prepare sql statement
+        sql = sql_stmt.format(
+            input1_databasename=input1_databasename,
+            input2_databasename=input2_databasename,
+        )
+
+        # Execute sql to be able to get the column names
+        cur = conn.cursor()
+        cur.execute(sql)
+        columns = [desc[0] for desc in cur.description]
+        conn.rollback()
+    except Exception as ex:
+        conn.rollback()
+        raise Exception(f"Error {ex} executing {sql}") from ex
+    finally:
+        conn.close()
+
+    return columns
+
+
 def create_table_as_sql(
     input1_path: Path,
     input1_layer: str,
@@ -329,6 +376,7 @@ def create_table_as_sql(
             columns_for_create = [
                 f'"{columnname}" {column_types[columnname]}\n'
                 for columnname in column_types
+                if columnname.lower() != "fid"
             ]
             sql = f"""
                 CREATE TABLE {output_databasename}."{output_layer}" (
@@ -372,12 +420,23 @@ def create_table_as_sql(
                 conn.execute(sql)
 
             # Insert data using the sql statement specified
-            columns_for_insert = [f'"{column[1]}"' for column in tmpcolumns]
-            sql = (
-                f'INSERT INTO {output_databasename}."{output_layer}" '
-                f'({", ".join(columns_for_insert)})\n{sql_stmt}'
-            )
-            conn.execute(sql)
+            try:
+                columns_for_insert = [f'"{column[1]}"' for column in tmpcolumns]
+                sql = (
+                    f'INSERT INTO {output_databasename}."{output_layer}" '
+                    f'({", ".join(columns_for_insert)})\n{sql_stmt}'
+                )
+                conn.execute(sql)
+            except Exception as ex:
+                ex_message = str(ex).lower()
+                if ex_message.startswith(
+                    "unique constraint failed:"
+                ) and ex_message.endswith(".fid"):
+                    ex.args = (
+                        f"{ex}: avoid this by not selecting or aliasing fid "
+                        '("select * will select fid!)',
+                    )
+                raise
 
             # Create spatial index if needed
             if create_spatial_index is True: